refactor: consolidate error utilities into utils/error
